Add files via upload
Changed the game update loop around so that the player data is transferred at the end of every update loop and the state checker is reset back to 0 at the end of every update loop. Will test with Eric to see if these changes imrpove the performance of the game.